+++ src/codec/GetParticipantIdResponseCodec.ts
@@ -7,7 +7,7 @@ import {GetParticipantIdResponse} from "../model/GetParticipantIdResponse";
 import {GetParticipantIdResponse as PbGetParticipantIdResponse} from "../generated/com/digitalasset/ledger/api/v1/admin/party_management_service_pb";
 
 export const GetParticipantIdResponseCodec: Codec<PbGetParticipantIdResponse, GetParticipantIdResponse> = {
-    deserialize(message: PbGetParticipantIdResponse): GetParticipantIdResponse{
+    deserialize(message: PbGetParticipantIdResponse): GetParticipantIdResponse {
         const object: GetParticipantIdResponse = {
             participantId: message.getParticipantId()
         };