+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Changed
+- support SDK 0.13.55 [#92](https://github.com/digital-asset/daml-js/pull/92)
+
+### Removed
+- **BREAKING**: `ExercisedEvent` no longer exposes the field `contractCreatingEventId` as it's been dropped by the SDK.
+
 ## [0.8.3]
 - fix bug to expose package management client from SDK 0.12.25 [#88](https://github.com/digital-asset/daml-js/pull/88)